+++ test/ExportSpec.hs
@@ -111,7 +111,7 @@ buildFixture comp = do
             "GLO"
             LocationDeclared
             "kg"
-            [ TechnosphereExchange prodU 1.0 unitU ReferenceProduct UUID.nil Nothing "" Nothing Nothing Nothing M.empty
+            [ TechnosphereExchange prodU 1.0 unitU ReferenceProduct UUID.nil Nothing "" Nothing Nothing Nothing M.empty noProperties
             , BiosphereExchange co2U 0.5 unitU Emission "" Nothing Nothing
             ]
             M.empty